Fix error

+++ packages/orchestrator/internal/sandbox/block/iov.go
@@ -11,11 +11,11 @@ import (
 // IOV_MAX is the limit of the vectors that can be passed in a single ioctl call.
 var IOV_MAX = utils.Must(getIOVMax())
 
-func getIOVMax() (int64, error) {
+func getIOVMax() (int, error) {
 	iovMax, err := sysconf.Sysconf(sysconf.SC_IOV_MAX)
 	if err != nil {
 		return 0, fmt.Errorf("failed to get IOV_MAX: %w", err)
 	}
 
-	return iovMax, nil
+	return int(iovMax), nil
 }
